+++ TAS.Common/Interfaces/IIngestDirectory.cs
@@ -12,14 +12,14 @@ namespace TAS.Common.Interfaces
     public interface IIngestDirectoryProperties : IMediaDirectoryProperties
     {
         TAspectConversion AspectConversion { get; }
-        decimal AudioVolume { get;  }
+        double AudioVolume { get;  }
         bool DeleteSource { get; }
         bool IsExport { get; }
         bool IsImport { get; }
         TVideoCodec VideoCodec { get; }
         TAudioCodec AudioCodec { get; }
-        decimal VideoBitrateRatio { get; }
-        decimal AudioBitrateRatio { get; }
+        double VideoBitrateRatio { get; }
+        double AudioBitrateRatio { get; }
         string EncodeParams { get; }
         TMovieContainerFormat ExportContainerFormat { get; }
         string ExportParams { get;  }
